+++ CHANGES.md
@@ -1,6 +1,8 @@
 # 3.4.8
 
 * `run-specs` now accepts string keys in addition to keyword keys
+* js runner executable may use `speclj.run.standard.arm()` instead of `speclj.run.standard.armed = true`
+* exports functions used by JavaScript for executing ClojureScript specs
 
 # 3.4.7